refactor: move prior helper functions to separate files.
